+++ src/vecrw.hpp
@@ -20,15 +20,16 @@
 
 #include "datarw.hpp"
 
-#include <cstdint>
 #include <vector>
+
+#include <stdint.h>
 
 
 class VecRW : public DataRW
 {
 private:
   std::vector<uint8_t> *_vec;
-  size_t                _idx;
+  uint64_t              _idx;
 
 public:
   void reset(std::vector<uint8_t> *vec);
@@ -39,10 +40,10 @@ public:
 public:
   bool   eof() const;
   bool   error() const;
-  size_t tell() const;
-  void   seek(const size_t idx);
+  uint64_t tell() const;
+  void   seek(const uint64_t idx);
 
 protected:
-  size_t _r(void *p, const size_t count);
-  size_t _w(const void *p, const size_t count);
+  uint64_t _r(void *p, const uint64_t count);
+  uint64_t _w(const void *p, const uint64_t count);
 };
